OP-451-refactored - OrderItem entity changed to view
